Ajout de modification et confirmation suppression, il reste contrainte 30j et garder parametre pour vencle

+++ rentmanager/rentmanager/src/main/java/com/epf/rentmanager/servlet/ClientListServlet.java
@@ -45,21 +45,31 @@ public class ClientListServlet extends HttpServlet {
         String action = request.getParameter("action");
         int clientID = Integer.parseInt(request.getParameter("clientID"));
         if (Objects.equals(action, "delete_client")) {
-            String prenom = request.getParameter("prenom");
-            String nom = request.getParameter("nom");
-            try {
-                clientService.delete(clientService.findById(clientID));
-                request.setAttribute("successMessage", "La suppression du client : " + prenom + " " + nom + " a été effectuée avec succès !");
-                response.sendRedirect(request.getContextPath() + "/users/list");
-            } catch (ServiceException e) {
-                response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Une erreur s'est produite lors de la suppression du véhicule.");
-                throw new RuntimeException(e);
+            String deleteornot = request.getParameter("deleteornot");
+            if (Objects.equals(deleteornot, "true")) {
+                String prenom = request.getParameter("prenom");
+                String nom = request.getParameter("nom");
+                try {
+                    clientService.delete(clientService.findById(clientID));
+                    request.setAttribute("successMessage", "La suppression du client : " + prenom + " " + nom + " a été effectuée avec succès !");
+                    response.sendRedirect(request.getContextPath() + "/users/list");
+                } catch (ServiceException e) {
+                    response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Une erreur s'est produite lors de la suppression du véhicule.");
+                    throw new RuntimeException(e);
+                }
+            }
+            else {
+                System.out.println("false");
+                doGet(request,response);
             }
         }
         if (Objects.equals(action, "modif_client")) {
-           
-        }
+            System.out.println("modif_client");
 
+        }
+        if (Objects.equals(action, "dont_delete_client")) {
+            doGet(request,response);
+        }
 
 
     }
