+++ rentmanager/rentmanager/src/main/java/com/epf/rentmanager/servlet/VehicleListServlet.java
@@ -2,6 +2,7 @@ package com.epf.rentmanager.servlet;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -10,7 +11,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.epf.rentmanager.exception.ServiceException;
+import com.epf.rentmanager.models.Reservation;
 import com.epf.rentmanager.models.Vehicle;
+import com.epf.rentmanager.service.ReservationService;
 import com.epf.rentmanager.service.VehicleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.context.support.SpringBeanAutowiringSupport;
@@ -22,6 +25,9 @@ public class VehicleListServlet extends HttpServlet {
     @Autowired
     private VehicleService vehicleService;
 
+    @Autowired
+    private ReservationService reservationService;
+
     @Override
     public void init() throws ServletException {
         super.init();
@@ -32,27 +38,41 @@ public class VehicleListServlet extends HttpServlet {
             throws ServletException, IOException {
         try {
             List<Vehicle> vehicles = vehicleService.findAll();
-            System.out.println(vehicles);
             request.setAttribute("vehicles", vehicles);
-        }
-        catch (Exception e) {
+            List<Reservation> resas = reservationService.findAll();
+            request.setAttribute("resas", resas);
+        } catch (Exception e) {
             e.printStackTrace();
         }
         this.getServletContext().getRequestDispatcher("/WEB-INF/views/vehicles/list.jsp").forward(request, response);
     }
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        int vehicleId = Integer.parseInt(request.getParameter("vehicleId"));
-        String constructeur = request.getParameter("constructeur");
-        String model = request.getParameter("model");
-        System.out.println(constructeur);
 
-        try {
-            vehicleService.delete(vehicleService.findById(vehicleId));
-            request.setAttribute("successMessage", "La suppression du vehicule : "+constructeur+" "+model+" a été effectuée avec succès !");
-            response.sendRedirect(request.getContextPath() + "/vehicles/list");
-        } catch (ServiceException e) {
-            response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Une erreur s'est produite lors de la suppression du véhicule.");
-            throw new RuntimeException(e);
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        String deleteornot = request.getParameter("deleteornot");
+        System.out.println(deleteornot);
+        if (Objects.equals(deleteornot, "true")) {
+            int vehicleId = Integer.parseInt(request.getParameter("vehicleId"));
+            String constructeur = request.getParameter("constructeur");
+            String model = request.getParameter("model");
+            System.out.println(constructeur);
+            try {
+                List<Reservation> reservations = reservationService.findResaByVehicleId(vehicleId);
+                int nb_reservations = reservations.size();
+            } catch (ServiceException e) {
+                throw new RuntimeException(e);
+            }
+            try {
+                vehicleService.delete(vehicleService.findById(vehicleId));
+                request.setAttribute("successMessage", "La suppression du vehicule : " + constructeur + " " + model + " a été effectuée avec succès !");
+                response.sendRedirect(request.getContextPath() + "/vehicles/list");
+            } catch (ServiceException e) {
+                response.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR, "Une erreur s'est produite lors de la suppression du véhicule.");
+                throw new RuntimeException(e);
+            }
+        }
+        else {
+            System.out.println("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA");
+            doGet(request,response);
         }
     }
 }
